Add mkiocccentry -m make option
Updated find_utils() to check for make(1) (default paths /usr/bin/make and
if not that /bin/make). New option to mkiocccentry is -m make to do this.
This will be needed for part of #1070.

Updated mkiocccentry(1) or the new -m make option.

Updated MKIOCCCENTRY_VERSION to "1.2.16 2025-02-06".
Updated SOUP_VERSION to "1.1.16 2025-02-06".
Updated TXZCHK_VERSION to "1.1.11 2025-02-06".