Disable ioccc_test.sh -m in test_ioccc/Makefile
Unfortunately this will not work after all because if MAKE is not set to
a path mkiocccentry will be given a path that does not exist so if one
has a system where the default make(1) path is not GNU compatible they
will have to run the script manually. This should fix the workflow
errors that have happened with the last few commits.