Bug fixes in mkiocccentry
Check for NULL or empty workdir and topdir prior to doing anything.

Change mkiocccentry arg count check to be exact as we no longer use extra args
(though it didn't hurt to have them it might be better to be clear to the user).
This required updating mkiocccentry_test.sh as well.

Updated MKIOCCCENTRY_VERSION to "1.2.29 2025-02-20".
Updated MKIOCCCENTRY_TEST_VERSION to "1.0.12 2025-02-20".