Make mkiocccentry use chkentry one time
Change mkiocccentry to form .auth.json and then .info.json and then run
chkentry(1) on both of the files. This will changed to running it with a
single arg - the topdir - but this change allows it to be easier to do once
chkentry(1) has been modified to do this (part of it was done - namely that
the topdir arg is there and 'set' but it's not used yet).

There was a sequenced exit code change in jparse/util.c. This will be
committed 'over there' but this is not strictly a sync as it was from
make seqcexit here.