Improve display of mkiocccentry_test.sh
It can be a bit confusing with all the different tests that it runs
(running it numerous times where it must fail and times where it must
succeed) so I added a '--' between invocations.

Previously it looked weird in that it would show that all warnings are
ignored and then immediately give an error (because the depth is too
deep). But before that warning it would show that one (because of 'test'
mode) could not use the tarball as a submission. So this should help
separate it a bit.