Make default path of tools in repo again
Due to some issues (like GitHub workflows) we cannot unfortunately have
the tool defaults in the tools. The scripts will have to be updated to
use the correct options if that is desired (and maybe it's not a good
idea for development as it means we have to install the tools each time
certain changes are made).