Add missing calls to errp() in txzchk
This might be important in some cases as we're now at the point of the
next contest. The calls are specific to the functions popen(3) in
pipe_open() and system(3) in shell_cmd(). In linux in some cases
popen() will set errno; in macOS it is unreliable. For system(3) it
appears that errno is not set so warnp()/errp() were changed to the
non-errno versions.

Updated TXZCHK_VERSION to "2.0.1 2025-03-02".