Resolve issues #1215 and #1207
Now the version checks for chkentry(1) are a >= check. Uses code from
jparse/verge.c. Its main() was moved to verge_main.c and verge.c now has
a new function vercmp(). 'verge.o' is linked into the library and
chkentry now uses it. Also there are new arrays that we called 'poisoned
versions' and this allows for bad versions to be excluded. Each tool and
some other versions have a minimum version allowed which at this time is
equivalent to the current release. If a version is incremented then the
minimum version would be changed to be the version at the time the
contest opens. In this way uploaded submissions will not be invalidated.
As for poisoned versions the lists are currently empty (just NULL
terminated - must be last element).

Issue #1207 was only resolved as it is an update to an array and is only
to help users out. This issue is what prompted issue #1215.

Updated gen_test_JSON.sh (under test_ioccc/) to have a new function -
get_version. This was necessary so we can use grep -v MIN_. If it was in
the same function it would cause another MIN_ macro to be excluded from
limit_ioccc.h which was a problem. Updated the script's version to
"1.0.2 2025-03-07".